Move rule related tests out of parser_test.go
